Update structrure from TOneV2_Dev_Backup
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C36490

+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE_StateBackup/TOneWhS_BE_Bkup/Tables/SalePriceList.sql
@@ -5,7 +5,9 @@
     [CurrencyID]    INT          NOT NULL,
     [EffectiveOn]   DATETIME     NULL,
     [SourceID]      VARCHAR (50) NULL,
-    [StateBackupID] INT          NOT NULL,
+    [StateBackupID] BIGINT       NOT NULL,
     CONSTRAINT [PK_SalePriceList] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
